% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispFieldST.R
\name{DispFieldST}
\alias{DispFieldST}
\title{Displacement fields for spatiotemporal data}
\usage{
DispFieldST(inputstack1, lag1, factv1, facth1, restricted = FALSE)
}
\arguments{
\item{factv1}{an odd integer for the vertical dimension of subgrids}

\item{facth1}{an odd integer for the horizontal dimension of subgrids}

\item{restricted}{logical (TRUE or FALSE)}
}
\value{

}
\description{
Displacement fields for spatiotemporal data
}
