% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispFieldSTall.R
\name{DispFieldSTall}
\alias{DispFieldSTall}
\title{Diplacement fields for spatiotemporal data when time lag is unknown}
\usage{
DispFieldSTall(inputstack1, lagmax, factv1, facth1, restricted = FALSE)
}
\arguments{
\item{inputstack1}{a raster stack with each raster layer representing an
instance of time. The raster stack should be organized such that the first
raster in the stack is the first observed spatial dataset and time
progresses forward with the third dimension index of the raster stack. The
raster stack should contain only numeric values. Any NA value will be
converted to a zero}

\item{lagmax}{an integer representing the maximum time lag}

\item{factv1}{an odd integer for the vertical dimension of subgrids}

\item{facth1}{an odd integer for the horizontal dimension of subgrids}

\item{restricted}{logical (TRUE or FALSE)}
}
\value{
A data frame is returned with the following column names: rowcent,
colcent, frowmin, frowmax, fcolmin, fcolmax, centx, centy, dispx, and
dispy. The rowcent and colcent column names are the row and column indices
for the center of each sub-grid; frowmin and frowmax are the sub-grid
minimum and maximum row indices; fcolmin and fcolmax are the sub-grid
minimum and maximum column indices; centx and centy are the projected
coordinates of the centre of the subgrid derived from the raster input
files; dispx and dispy are the orthoganal velocity vectors in units of
space per timestep in the horizontal and vertical directions in the same
spatial units as the projected coordinates of the raster input files.
}
\description{
This is an implementation of a novel algorithm that differs from more
traditional digital image correlation implementations that are applied in the
\code{\link{DispField}} and \code{\link{DispFieldbb}} functions. The function
calculates a displacement field representing persistent movement based on the
cross-covariance in a raster stack (in this case a sequential series of
rasters) presumably representing spatial population abundance or density at
more than two different instances of time. If analysis is restricted to only
two time instances, \code{\link{DispField}} is more appropriate.
}
\details{
The DispFieldSTall function has the same inner workings as the
\code{\link{DispFieldST}} function except that instead of specifying a
specific time lag, the user specifies a maximum time lag. The function then
cycles through all possible lags up to the maximum time lag and choses the
for each location the maximum velocity.

Caution is warranted when defining the sub-grid dimensions because the
function can produce erroneous results when sub-grids are too small.
}
\examples{
Vec1 <- c(1:5, 0, 0, 0, 0)
Mat1 <- Vec1
for(i in 2:9) Mat1 <- rbind(Mat1, Vec1)
Mat1

Vec2 <- c(0, 1:5, 0, 0, 0)
Mat2 <- Vec2
for(i in 2:9) Mat2 <- rbind(Mat2, Vec2)
Mat2

Vec3 <- c(0, 0, 1:5, 0, 0)
Mat3 <- Vec3
for(i in 2:9) Mat3 <- rbind(Mat3, Vec3)
Mat3

Vec4 <- c(0, 0, 0, 1:5, 0)
Mat4 <- Vec4
for(i in 2:9) Mat4 <- rbind(Mat4, Vec4)
Mat4

# Note that rasterizing a matrix causes it to be rotated 90 degrees.
# Therefore, any shift in the x direction is in fact now a shift in the y direction
rast1 <- terra::rast(Mat1)
terra::plot(rast1)
rast2 <- terra::rast(Mat2)
terra::plot(rast2)
rast3 <- terra::rast(Mat3)
terra::plot(rast3)
rast4 <- terra::rast(Mat4)
terra::plot(rast4)

teststack1 <- c(rast1, rast2, rast3, rast4)
(VFdf4 <- DispFieldSTall(teststack1, lagmax = 2, factv1 = 9, facth1 = 9))
# block is moving downward at a speed of 0.1111111 units of space per unit of time
# dispy = -0.1111111
}
