% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LBMfromncfpkg}
\alias{LBMfromncfpkg}
\title{Larch budmoth defoliation data from the ncf R package}
\format{
A data-frame with 135 rows and 40 columns. The first two are the x-
and y-coordinates (in m), the following 38 columns represent defoliation in
years 1961 through 1998.
}
\source{
\url{https://cran.r-project.org/web/packages/ncf/index.html}
}
\usage{
data(LBMfromncfpkg)

ICvectorfields::RastStackData(LBMfromncfpkg)
}
\description{
These are the data in Bjornstad et al. (2002). The data
are formatted such that they can easily be converted to
a raster stack using
ICvectorfields::RastStackData(LBMfromncfpkg).
}
\references{
Bjornstad, O.N., Peltonen, M., Liebhold, A.M., and Baltensweiler,
W. (2002) Waves of larch budmoth outbreaks in the European Alps. Science,
298, 1020-1023. \url{https://doi.org/10.1126/science.1075182}
}
\keyword{datasets}
