% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatternDetect.R
\name{PatternDetect}
\alias{PatternDetect}
\title{Detect Patterns in Vector Fields}
\usage{
PatternDetect(vfdf)
}
\arguments{
\item{vfdf}{A data frame as returned by \code{\link{DispField}},
\code{\link{DispFieldST}}, or \code{\link{DispFielSTall}} with at least
five rows (more is better)}
}
\value{
A data frame as returned by \code{\link{DispField}},
\code{\link{DispFieldST}}, or \code{\link{DispFielSTall}}, with an
additional column called Pattern in which the patterns around each focal
cell are categorized as convergence, divergence, wave or NA
}
\description{
Detect patterns in vector fields represented on a grid by looking in the
rook's neighbourhood of each grid cell. Three patterns are detected: convergences
occur when the vectors in the four adjacent cells point towards the focal
cell; divergences occur when the vectors in the four adjacent cells point
away from the focal cell; waves occur when directions of vectors are
consistent across all of the adjacent cells as well as the focal cell. In the
wave pattern vectors can point up and right, up and left, down and right, or
down and left.
}
\examples{
# creating convergence/divergence patterns
Mat1 <- matrix(rep(0,9*9), nrow = 9)
Mat1[3, c(4, 6)] <- 1
Mat1[7, c(4, 6)] <- 1
Mat1[c(4, 6), 3] <- 1
Mat1[c(4, 6), 7] <- 1
Mat1

Mat2 <- matrix(rep(0,9*9), nrow = 9)
Mat2[2, c(4, 6)] <- 1
Mat2[8, c(4, 6)] <- 1
Mat2[c(4, 6), 2] <- 1
Mat2[c(4, 6), 8] <- 1
Mat2

# rasterizing
rast1 <- terra::rast(Mat1)
terra::plot(rast1)
rast2 <- terra::rast(Mat2)
terra::plot(rast2)

# Detecting a divergence
(VFdf1 <- DispField(rast1, rast2, factv1 = 3, facth1 = 3, restricted = TRUE))
(patdf1 <- PatternDetect(VFdf1))

# Detecting a convergence
(VFdf2 <- DispField(rast2, rast1, factv1 = 3, facth1 = 3, restricted = TRUE))
(patdf2 <- PatternDetect(VFdf2))
}
