% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReflMat.R
\name{ReflMat}
\alias{ReflMat}
\title{ReflMat}
\usage{
ReflMat(dim1)
}
\arguments{
\item{dim1}{is the dimension (number of rows and columns) of the output matrix}
}
\value{
numeric matrix
}
\description{
This function creates a reflection matrix with ones along the diagonal that traverses from the lower
left corner to the upper right corner of the square matrix of dimension dim1 X dim1.
}
\examples{
ReflMat(5)
}
