% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgridMoransI.R
\name{SubgridMoransI}
\alias{SubgridMoransI}
\title{Compute Moran's I for sub-grids}
\usage{
SubgridMoransI(inputrast1, factv1, facth1, rad1 = 1)
}
\arguments{
\item{inputrast1}{a raster as produced by terra::rast}

\item{factv1}{an odd integer for the vertical dimension of sub-grids}

\item{facth1}{an odd integer for the horizontal dimension of sub-grids}

\item{rad1}{an number indicating the neighbourhood radius in rows/columns.
Any cell within a distance of rad1 cells of the focal cell is considered to
be in its neighbourhood.}
}
\value{
A data frame is returned with the following column names: rowcent,
colcent, frowmin, frowmax, fcolmin, fcolmax, and a column for the output
Moran's I statistic.
}
\description{
A function that facilitates calculation of Moran's I at the sub-grid level.
This may be useful for evaluating spatial autocorrelation as a driver  of
movement speed or direction if used in tandem with \code{\link{DispField}},
\code{\link{DispFieldST}}, or \code{\link{DispFieldSTall}}.
}
\details{
Note that when using radius to define the neighbourhood, a radius of one
corresponds to the rook's neibhourhood, whereas a radius of 1.5 corresponds
to the queen's neighbourhood.
}
\examples{
(TestMat <- matrix(c(1, 0, 1, 0, 1,
                     0, 1, 0, 1, 0,
                     1, 0, 1, 0, 1,
                     0, 1, 0, 1, 0,
                     1, 0, 1, 0, 1),
                    nrow = 5))

TestRast <- terra::rast(TestMat)
terra::plot(TestRast)

SubgridMoransI(TestRast, factv1 = 5, facth1 = 5, rad1 = 1)
# using rad1 = 1 is equivalent to using the rooks neighbourhood
# and so the output should be -1.

}
