% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgridStats.R
\name{SubgridStats}
\alias{SubgridStats}
\title{Compute statistics for subgrids.}
\usage{
SubgridStats(inputrast1, factv1, facth1, statistic = "var")
}
\arguments{
\item{inputrast1}{a raster as produced by terra::rast}

\item{factv1}{an odd integer for the vertical dimension of sub-grids}

\item{facth1}{an odd integer for the horizontal dimension of sub-grids}

\item{statistic}{desired output statistic: It should be one of "mean",
"var", or "sum". Default setting is var.}
}
\value{
A data frame is returned with the following column names: rowcent,
colcent, frowmin, frowmax, fcolmin, fcolmax, and a column for the output
statistic.
}
\description{
A function that facilitates calculation of statistics at the sub-grid level.
This may be useful for computing predictors of movement speed or direction if
used in tandem with \code{\link{DispField}}, \code{\link{DispFieldST}}, or
\code{\link{DispFieldSTall}}.
}
\examples{
(TestMat <- matrix(c(1, 0, 1, 0, 1,
                    0, 1, 0, 1, 0,
                    1, 0, 1, 0, 1,
                    0, 1, 0, 1, 0,
                    1, 0, 1, 0, 1),
                    nrow = 5))

TestRast <- terra::rast(TestMat)
terra::plot(TestRast)

SubgridStats(TestRast, factv1 = 5, facth1 = 5, statistic = "mean")
SubgridStats(TestRast, factv1 = 5, facth1 = 5, statistic = "var")
SubgridStats(TestRast, factv1 = 5, facth1 = 5, statistic = "sum")

}
